src/requirements.txt
python-telegram-bot
uxly-1shot-client
pydantic